+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+### 🌟 New instrumentation
+
+- Add `captureTemplate` and `captureArguments` options to the log4j, java-util-logging, and
+  jboss-logmanager logging instrumentations, capturing the log message template and arguments as
+  separate `log.body.template` / `log.body.parameters` attributes. This extends the same option
+  that was previously added for logback in
+  [#15423](https://github.com/open-telemetry/opentelemetry-java-instrumentation/pull/15423).
+
 ## Version 2.29.0 (2026-06-19)
 
 This release targets the OpenTelemetry SDK 1.63.0.